requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-pptx==0.6.21
python-multipart==0.0.9
email-validator==2.2.0